Handle additional search parameters on webinars page
